refactor: some syntax + if conditions
