+++ packages/app/app/services/message/moderation/AdminActionColorMap.ts
@@ -1,10 +1,10 @@
 import { AdminActionType } from "@esposter/db-schema";
 
-export const AdminActionColorMap: Record<AdminActionType, string> = {
+export const AdminActionColorMap = {
   [AdminActionType.BanUser]: "error",
   [AdminActionType.ForceMute]: "warning",
   [AdminActionType.ForceUnmute]: "success",
   [AdminActionType.KickFromRoom]: "error",
   [AdminActionType.KickFromVoice]: "warning",
   [AdminActionType.TimeoutUser]: "warning",
-};
+} as const satisfies Record<AdminActionType, string>;
