+++ packages/app/server/trpc/plugins/timeoutPlugin.ts
@@ -1,6 +1,6 @@
 import type { AuthedContext } from "@@/server/models/auth/AuthedContext";
 
-import { assertNotTimedOut } from "@@/server/services/moderation/assertNotTimedOut";
+import { assertNotTimedOut } from "@@/server/services/message/moderation/assertNotTimedOut";
 import { selectRoomSchema } from "@esposter/db-schema";
 import { initTRPC } from "@trpc/server";
 import { z } from "zod";
