+++ packages/app/server/trpc/routers/index.ts
@@ -11,8 +11,8 @@ import { friendRouter } from "@@/server/trpc/routers/friend";
 import { friendRequestRouter } from "@@/server/trpc/routers/friendRequest";
 import { likeRouter } from "@@/server/trpc/routers/like";
 import { messageRouter } from "@@/server/trpc/routers/message";
-import { moderationRouter } from "@@/server/trpc/routers/moderation";
 import { emojiRouter } from "@@/server/trpc/routers/message/emoji";
+import { moderationRouter } from "@@/server/trpc/routers/message/moderation";
 import { postRouter } from "@@/server/trpc/routers/post";
 import { pushSubscriptionRouter } from "@@/server/trpc/routers/pushSubscription";
 import { roleRouter } from "@@/server/trpc/routers/role";
