+++ packages/app/server/trpc/routers/message/index.ts
@@ -24,10 +24,10 @@ import { on } from "@@/server/services/events/on";
 import { messageEventEmitter } from "@@/server/services/message/events/messageEventEmitter";
 import { roomEventEmitter } from "@@/server/services/message/events/roomEventEmitter";
 import { isRoomId } from "@@/server/services/message/isRoomId";
+import { assertNotTimedOut } from "@@/server/services/message/moderation/assertNotTimedOut";
 import { readMessages } from "@@/server/services/message/readMessages";
 import { searchMessages } from "@@/server/services/message/searchMessages";
 import { updateMessage } from "@@/server/services/message/updateMessage";
-import { assertNotTimedOut } from "@@/server/services/moderation/assertNotTimedOut";
 import { router } from "@@/server/trpc";
 import { isMember } from "@@/server/trpc/middleware/userToRoom/isMember";
 import { timeoutPlugin } from "@@/server/trpc/plugins/timeoutPlugin";
